refactor: order 도메인 폴더 분리
