+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "prisma.pinToPrisma6": true
+}
