+++ src/account/account.service.ts
@@ -7,7 +7,7 @@ import { TransferAccountBalanceDto } from './dto/transfer-account-balance.dto';
 import { DepositAccountBalanceDto } from './dto/deposit-account-balance.dto copy';
 import { WithdrawStockDto } from './dto/withdraw-stock.dto';
 import { DepositStockDto } from './dto/deposit-stock.dto';
-import { OrderException } from 'src/stock/order/error/order.exception';
+import { OrderException } from 'src/order/error/order.exception';
 import { StockException } from 'src/stock/error/stock.exception';
 
 @Injectable()
